fix Test_MTU
